Add `SIMDVector::clamp_min` and use it in activations
For clamping we do not need the NaN-propagation of SIMDVector::max.  One of the
operands is under our control and therefore not NaN, so we can use SIMD
instructions such that NaN propagation is only applied for the clamped values.

E.g. for AVX this means that we can rely on `_mm256_max_ps`'s behavior to copy
the second operand in the presence of NaN.