Replace reduce_generic function by macro
Since the reduce_generic function does not have a target_feature attribute, we
often end up with loops where the intrinsics are not inlined. Replace the
function by a macro, so that the loop is expanded into a method with
target_feature and we get vectorized loops with inlined intrinstics.